Fix type of color components.
[FIXED]     Color components are scanned as floats, not CGFloats.

+++ UIColor+Expanded.m
@@ -589,7 +589,7 @@ static NSMutableDictionary *_CrayolaNameCache = nil;
 	}
 	
 	const NSUInteger kMaxComponents = 4;
-	CGFloat c[kMaxComponents];
+	float c[kMaxComponents];
 	NSUInteger i = 0;
 	
 	if (![scanner scanFloat:&c[i++]]) {
